update mocks and dht tests
-reorganize dht tests
