+++ lbrynet/tests/functional/dht/test_bootstrap_network.py
@@ -1,0 +1,10 @@
+from dht_test_environment import TestKademliaBase
+
+
+class TestKademliaBootstrap(TestKademliaBase):
+    """
+    Test initializing the network / connecting the seed nodes
+    """
+
+    def test_bootstrap_seed_nodes(self):
+        pass
