+++ lbrynet/tests/mocks.py
@@ -1,21 +1,18 @@
 import base64
-import struct
 import io
 
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.hazmat.primitives import serialization
-from twisted.internet import defer, error
+from twisted.internet import defer
 from twisted.python.failure import Failure
 
 from lbrynet.core.client.ClientRequest import ClientRequest
 from lbrynet.core.Error import RequestCanceledError
 from lbrynet.core import BlobAvailability
-from lbrynet.core.utils import generate_id
 from lbrynet.dht.node import Node as RealNode
 from lbrynet.daemon import ExchangeRateManager as ERM
 from lbrynet import conf
-from util import debug_kademlia_packet
 
 KB = 2**10
 PUBLIC_EXPONENT = 65537  # http://www.daemonology.net/blog/2009-06-11-cryptographic-right-answers.html
@@ -40,6 +37,9 @@ class Node(RealNode):
 
     def stop(self):
         return defer.succeed(None)
+
+    def start(self, known_node_addresses=None):
+        return self.joinNetwork(known_node_addresses)
 
 
 class FakeNetwork(object):
@@ -188,8 +188,14 @@ class Wallet(object):
     def get_info_exchanger(self):
         return PointTraderKeyExchanger(self)
 
+    def update_peer_address(self, peer, address):
+        pass
+
     def get_wallet_info_query_handler_factory(self):
         return PointTraderKeyQueryHandlerFactory(self)
+
+    def get_unused_address_for_peer(self, peer):
+        return defer.succeed("bDtL6qriyimxz71DSYjojTBsm6cpM1bqmj")
 
     def reserve_points(self, *args):
         return True
@@ -250,17 +256,11 @@ class Announcer(object):
     def immediate_announce(self, *args):
         pass
 
-    def run_manage_loop(self):
-        pass
-
     def start(self):
         pass
 
     def stop(self):
         pass
-
-    def get_next_announce_time(self):
-        return 0
 
 
 class GenFile(io.RawIOBase):
@@ -410,89 +410,3 @@ def mock_conf_settings(obj, settings={}):
         conf.settings = original_settings
 
     obj.addCleanup(_reset_settings)
-
-
-MOCK_DHT_NODES = [
-    "000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-    "FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF",
-    "DEADBEEFDEADBEEFDEADBEEFDEADBEEFDEADBEEFDEADBEEFDEADBEEFDEADBEEFDEADBEEFDEADBEEFDEADBEEFDEADBEEF",
-]
-
-MOCK_DHT_SEED_DNS = { # these map to mock nodes 0, 1, and 2
-    "lbrynet1.lbry.io": "10.42.42.1",
-    "lbrynet2.lbry.io": "10.42.42.2",
-    "lbrynet3.lbry.io": "10.42.42.3",
-}
-
-
-def resolve(name, timeout=(1, 3, 11, 45)):
-    if name not in MOCK_DHT_SEED_DNS:
-        return defer.fail(error.DNSLookupError(name))
-    return defer.succeed(MOCK_DHT_SEED_DNS[name])
-
-
-class MockUDPTransport(object):
-    def __init__(self, address, port, max_packet_size, protocol):
-        self.address = address
-        self.port = port
-        self.max_packet_size = max_packet_size
-        self._node = protocol._node
-
-    def write(self, data, address):
-        dest = MockNetwork.peers[address][0]
-        debug_kademlia_packet(data, (self.address, self.port), address, self._node)
-        dest.datagramReceived(data, (self.address, self.port))
-
-
-class MockUDPPort(object):
-    def __init__(self, protocol):
-        self.protocol = protocol
-
-    def startListening(self, reason=None):
-        return self.protocol.startProtocol()
-
-    def stopListening(self, reason=None):
-        return self.protocol.stopProtocol()
-
-
-class MockNetwork(object):
-    peers = {}  # (interface, port): (protocol, max_packet_size)
-
-    @classmethod
-    def add_peer(cls, port, protocol, interface, maxPacketSize):
-        interface = protocol._node.externalIP
-        protocol.transport = MockUDPTransport(interface, port, maxPacketSize, protocol)
-        cls.peers[(interface, port)] = (protocol, maxPacketSize)
-
-
-def listenUDP(port, protocol, interface='', maxPacketSize=8192):
-    MockNetwork.add_peer(port, protocol, interface, maxPacketSize)
-    return MockUDPPort(protocol)
-
-
-def address_generator(address=(10, 42, 42, 1)):
-    def increment(addr):
-        value = struct.unpack("I", "".join([chr(x) for x in list(addr)[::-1]]))[0] + 1
-        new_addr = []
-        for i in range(4):
-            new_addr.append(value % 256)
-            value >>= 8
-        return tuple(new_addr[::-1])
-
-    while True:
-        yield "{}.{}.{}.{}".format(*address)
-        address = increment(address)
-
-
-def mock_node_generator(count=None, mock_node_ids=MOCK_DHT_NODES):
-    if mock_node_ids is None:
-        mock_node_ids = MOCK_DHT_NODES
-
-    for num, node_ip in enumerate(address_generator()):
-        if count and num >= count:
-            break
-        if num >= len(mock_node_ids):
-            node_id = generate_id().encode('hex')
-        else:
-            node_id = mock_node_ids[num]
-        yield (node_id, node_ip)
