+++ lbrynet/tests/unit/dht/test_encoding.py
@@ -4,9 +4,9 @@
 # the GNU Lesser General Public License Version 3, or any later version.
 # See the COPYING file included in this archive
 
-import unittest
-
+from twisted.trial import unittest
 import lbrynet.dht.encoding
+
 
 class BencodeTest(unittest.TestCase):
     """ Basic tests case for the Bencode implementation """
@@ -16,7 +16,7 @@ class BencodeTest(unittest.TestCase):
         self.cases = ((42, 'i42e'),
                       ('spam', '4:spam'),
                       (['spam', 42], 'l4:spami42ee'),
-                      ({'foo':42, 'bar':'spam'}, 'd3:bar4:spam3:fooi42ee'),
+                      ({'foo': 42, 'bar': 'spam'}, 'd3:bar4:spam3:fooi42ee'),
                       # ...and now the "real life" tests
                       ([['abc', '127.0.0.1', 1919], ['def', '127.0.0.1', 1921]],
                        'll3:abc9:127.0.0.1i1919eel3:def9:127.0.0.1i1921eee'))
@@ -45,12 +45,3 @@ class BencodeTest(unittest.TestCase):
         for encodedValue in self.badDecoderCases:
             self.failUnlessRaises(
                 lbrynet.dht.encoding.DecodeError, self.encoding.decode, encodedValue)
-
-def suite():
-    suite = unittest.TestSuite()
-    suite.addTest(unittest.makeSuite(BencodeTest))
-    return suite
-
-if __name__ == '__main__':
-    # If this module is executed from the commandline, run all its tests
-    unittest.TextTestRunner().run(suite())
