+++ lbrynet/tests/unit/dht/test_kbucket.py
@@ -4,23 +4,41 @@
 # the GNU Lesser General Public License Version 3, or any later version.
 # See the COPYING file included in this archive
 
-import unittest
-
+from twisted.trial import unittest
+import struct
+from lbrynet.core.utils import generate_id
 from lbrynet.dht import kbucket
-import lbrynet.dht.contact as contact
+from lbrynet.dht.contact import ContactManager
 from lbrynet.dht import constants
+
+
+def address_generator(address=(10, 42, 42, 1)):
+    def increment(addr):
+        value = struct.unpack("I", "".join([chr(x) for x in list(addr)[::-1]]))[0] + 1
+        new_addr = []
+        for i in range(4):
+            new_addr.append(value % 256)
+            value >>= 8
+        return tuple(new_addr[::-1])
+
+    while True:
+        yield "{}.{}.{}.{}".format(*address)
+        address = increment(address)
+
 
 class KBucketTest(unittest.TestCase):
     """ Test case for the KBucket class """
     def setUp(self):
-        self.kbucket = kbucket.KBucket(0, 2**160)
+        self.address_generator = address_generator()
+        self.contact_manager = ContactManager()
+        self.kbucket = kbucket.KBucket(0, 2**constants.key_bits, generate_id())
 
     def testAddContact(self):
         """ Tests if the bucket handles contact additions/updates correctly """
         # Test if contacts can be added to empty list
         # Add k contacts to bucket
         for i in range(constants.k):
-            tmpContact = contact.Contact('tempContactID%d' % i, str(i), i, i)
+            tmpContact = self.contact_manager.make_contact(generate_id(), next(self.address_generator), 4444, 0, None)
             self.kbucket.addContact(tmpContact)
             self.failUnlessEqual(
                 self.kbucket._contacts[i],
@@ -28,8 +46,7 @@ class KBucketTest(unittest.TestCase):
                 "Contact in position %d not the same as the newly-added contact" % i)
 
         # Test if contact is not added to full list
-        i += 1
-        tmpContact = contact.Contact('tempContactID%d' % i, str(i), i, i)
+        tmpContact = self.contact_manager.make_contact(generate_id(), next(self.address_generator), 4444, 0, None)
         self.failUnlessRaises(kbucket.BucketFull, self.kbucket.addContact, tmpContact)
 
         # Test if an existing contact is updated correctly if added again
@@ -48,14 +65,17 @@ class KBucketTest(unittest.TestCase):
 
 
         # Add k-2 contacts
+        node_ids = []
         if constants.k >= 2:
             for i in range(constants.k-2):
-                tmpContact = contact.Contact(i, i, i, i)
+                node_ids.append(generate_id())
+                tmpContact = self.contact_manager.make_contact(node_ids[-1], next(self.address_generator), 4444, 0, None)
                 self.kbucket.addContact(tmpContact)
         else:
             # add k contacts
             for i in range(constants.k):
-                tmpContact = contact.Contact(i, i, i, i)
+                node_ids.append(generate_id())
+                tmpContact = self.contact_manager.make_contact(node_ids[-1], next(self.address_generator), 4444, 0, None)
                 self.kbucket.addContact(tmpContact)
 
         # try to get too many contacts
@@ -65,8 +85,8 @@ class KBucketTest(unittest.TestCase):
                         'Returned list should not have more than k entries!')
 
         # verify returned contacts in list
-        for i in range(constants.k-2):
-            self.failIf(self.kbucket._contacts[i].id != i,
+        for node_id, i in zip(node_ids, range(constants.k-2)):
+            self.failIf(self.kbucket._contacts[i].id != node_id,
                         "Contact in position %s not same as added contact" % (str(i)))
 
         # try to get too many contacts
@@ -89,25 +109,15 @@ class KBucketTest(unittest.TestCase):
 
     def testRemoveContact(self):
         # try remove contact from empty list
-        rmContact = contact.Contact('TestContactID1', '127.0.0.1', 1, 1)
+        rmContact = self.contact_manager.make_contact(generate_id(), next(self.address_generator), 4444, 0, None)
         self.failUnlessRaises(ValueError, self.kbucket.removeContact, rmContact)
 
         # Add couple contacts
         for i in range(constants.k-2):
-            tmpContact = contact.Contact('tmpTestContactID%d' % i, str(i), i, i)
+            tmpContact = self.contact_manager.make_contact(generate_id(), next(self.address_generator), 4444, 0, None)
             self.kbucket.addContact(tmpContact)
 
         # try remove contact from empty list
         self.kbucket.addContact(rmContact)
         result = self.kbucket.removeContact(rmContact)
         self.failIf(rmContact in self.kbucket._contacts, "Could not remove contact from bucket")
-
-
-def suite():
-    suite = unittest.TestSuite()
-    suite.addTest(unittest.makeSuite(KBucketTest))
-    return suite
-
-if __name__ == '__main__':
-    # If this module is executed from the commandline, run all its tests
-    unittest.TextTestRunner().run(suite())
