+++ lbrynet/tests/util.py
@@ -5,19 +5,10 @@ import os
 import tempfile
 import shutil
 import mock
-import logging
 
-from lbrynet.dht.encoding import Bencode
-from lbrynet.dht.error import DecodeError
-from lbrynet.dht.msgformat import DefaultFormat
-from lbrynet.dht.msgtypes import ResponseMessage, RequestMessage, ErrorMessage
 
-_encode = Bencode()
-_datagram_formatter = DefaultFormat()
 DEFAULT_TIMESTAMP = datetime.datetime(2016, 1, 1)
 DEFAULT_ISO_TIME = time.mktime(DEFAULT_TIMESTAMP.timetuple())
-
-log = logging.getLogger("lbrynet.tests.util")
 
 
 def mk_db_and_blob_dir():
@@ -49,28 +40,5 @@ def resetTime(test_case, timestamp=DEFAULT_TIMESTAMP):
     patcher.start().return_value = timestamp
     test_case.addCleanup(patcher.stop)
 
-
 def is_android():
     return 'ANDROID_ARGUMENT' in os.environ # detect Android using the Kivy way
-
-
-def debug_kademlia_packet(data, source, destination, node):
-    if log.level != logging.DEBUG:
-        return
-    try:
-        packet = _datagram_formatter.fromPrimitive(_encode.decode(data))
-        if isinstance(packet, RequestMessage):
-            log.debug("request %s --> %s %s (node time %s)", source[0], destination[0], packet.request,
-                      node.clock.seconds())
-        elif isinstance(packet, ResponseMessage):
-            if isinstance(packet.response, (str, unicode)):
-                log.debug("response %s <-- %s %s (node time %s)", destination[0], source[0], packet.response,
-                          node.clock.seconds())
-            else:
-                log.debug("response %s <-- %s %i contacts (node time %s)", destination[0], source[0],
-                          len(packet.response), node.clock.seconds())
-        elif isinstance(packet, ErrorMessage):
-            log.error("error %s <-- %s %s (node time %s)", destination[0], source[0], packet.exceptionType,
-                      node.clock.seconds())
-    except DecodeError:
-        log.exception("decode error %s --> %s (node time %s)", source[0], destination[0], node.clock.seconds())
